separating the pages and components logic from view
