--- src/pages/AddAnnouncement.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-import ManageAnnouncements from "../components/ManageAnnouncements";
-
-
-function AddAnnouncement() {
-  return (
-    <ManageAnnouncements info={null} />
-  )
-}
-
-export default AddAnnouncement;
