+++ src/pages/AddAnnouncement/AddAnnouncement.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+import ManageAnnouncements from "../../components/ManageAnnouncements/ManageAnnouncements.jsx";
+
+
+function AddAnnouncement() {
+  return (
+    <ManageAnnouncements info={null} />
+  )
+}
+
+export default AddAnnouncement;
